requirements.txt
pyyaml
opencv-contrib-python
matplotlib
numpy
tqdm
tensorboard
torchmetrics
segmentation-models-pytorch